+++ src/turn/static_service.rs
@@ -2,96 +2,38 @@
 //!
 //! [ICE]: https://webrtcglossary.com/ice/
 
-use std::str::FromStr;
-
 use async_trait::async_trait;
 use medea_client_api_proto::{IceServer, PeerId, RoomId};
 
 use crate::{
-    conf::turn::StaticCredentials,
+    conf::turn::RtcIceServer,
     turn::{IceUser, TurnAuthService, TurnServiceErr, UnreachablePolicy},
 };
-
-/// Kind of [`StaticIceUser`].
-#[derive(Clone, Copy, Debug)]
-enum Kind {
-    /// This is [TURN] [ICE] user.
-    ///
-    /// [TURN]: https://webrtcglossary.com/turn/
-    /// [ICE]: https://webrtcglossary.com/ice/
-    Turn,
-
-    /// This is [STUN] [ICE] user.
-    ///
-    /// [STUN]: https://webrtcglossary.com/stun/
-    /// [ICE]: https://webrtcglossary.com/ice/
-    Stun,
-}
-
-/// Error which indicates that incorrect [`Kind`] was found while parsing.
-#[derive(Debug)]
-pub struct InvalidKindErr {
-    /// TURN/STUN address on which this error happened.
-    pub address: String,
-
-    /// Found incorrect [`Kind`].
-    pub kind: String,
-}
-
-impl FromStr for Kind {
-    type Err = InvalidKindErr;
-
-    /// Lookups first 4 symbols and if the are `stun` or `turn`, then returns
-    /// matching [`Kind`].
-    ///
-    /// If incorrect symbols are found, then returns [`InvalidKindErr`].
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.len() < 4 {
-            return Err(InvalidKindErr {
-                address: s.to_string(),
-                kind: s.to_string(),
-            });
-        }
-
-        match &s[0..4] {
-            "stun" => Ok(Kind::Stun),
-            "turn" => Ok(Kind::Turn),
-            _ => Err(InvalidKindErr {
-                address: s.to_string(),
-                kind: s[0..3].to_string(),
-            }),
-        }
-    }
-}
 
 /// Static [ICE] user credentials which will be provided to the client.
 #[derive(Debug, Clone)]
 pub struct StaticIceUser {
-    /// Address of Turn server.
-    address: String,
+    /// URLs which can be used to connect to the server.
+    urls: Vec<String>,
 
-    /// Username for authorization.
+    /// Username to use during the authentication process.
     username: Option<String>,
 
-    /// Password for authorization.
-    pass: Option<String>,
-
-    /// Kind of this [`StaticIceUser`].
-    kind: Kind,
+    /// The credential to use when logging into the server.
+    credential: Option<String>,
 }
 
 impl StaticIceUser {
-    /// Returns new [`StaticIceUser`] with a provided [`StaticCredentials`].
+    /// Returns new [`StaticIceUser`] with the provided [`RtcIceServers`].
     ///
     /// Returns [`TurnServiceErr`] if incorrect [`Kind`] found in the provided
-    /// [`StaticCredentials`].
-    fn new(creds: StaticCredentials) -> Result<Self, TurnServiceErr> {
-        Ok(Self {
-            kind: creds.address.parse()?,
-            address: creds.address,
-            username: creds.username,
-            pass: creds.pass,
-        })
+    /// [`RtcIceServers`].
+    fn new(servers: RtcIceServer) -> Self {
+        Self {
+            urls: servers.urls,
+            username: servers.username,
+            credential: servers.credential,
+        }
     }
 }
 
@@ -99,9 +41,9 @@ impl StaticIceUser {
     /// Returns [`IceServer`] of this [`StaticIceUser`].
     pub fn ice_server(&self) -> IceServer {
         IceServer {
-            urls: vec![self.address.clone()],
+            urls: self.urls.clone(),
             username: self.username.clone(),
-            credential: self.pass.clone(),
+            credential: self.credential.clone(),
         }
     }
 }
@@ -115,15 +57,8 @@ pub struct StaticService(Vec<StaticIceUser>);
 impl StaticService {
     /// Returns new [`StaticService`] based on the provided
     /// [`conf::turn::Static`].
-    ///
-    /// Returns [`TurnServiceErr::InvalidKindInStaticCredentials`] if incorrect
-    /// [`Kind`] found in the provided [`StaticCredentials`].
-    pub fn new(creds: &[StaticCredentials]) -> Result<Self, TurnServiceErr> {
-        let mut ice_users = Vec::new();
-        for c in creds {
-            ice_users.push(StaticIceUser::new(c.clone())?);
-        }
-        Ok(Self(ice_users))
+    pub fn new(servers: Vec<RtcIceServer>) -> Self {
+        Self(servers.into_iter().map(StaticIceUser::new).collect())
     }
 }